Add tests for the MemoryIngestVersionManagerDao
